+++ packages/flame/CHANGELOG.md
@@ -56,6 +56,7 @@
  - `HasGameRef` can now operate independently from `Game`
  - `initialDelay` and `peakDelay` for effects to handle time before and after an effect
  - `component.onMount` now runs every time a component gets a new parent
+ - Add collision detection between child components
 
 ## [1.0.0-releasecandidate.13]
  - Fix camera not ending up in the correct position on long jumps